+++ src/containers/Layout/Styles.ts
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 
 export const Container = styled.div`
   display: grid;
-  grid-template-columns: 260px 1fr 320px;
+  grid-template-columns: 260px 1fr;
   height: 100vh;
   margin: auto;
   max-width: 1400px;
