+++ src/containers/LeftNav/Styles.ts
@@ -1,10 +1,31 @@
 import styled from 'styled-components';
 
-export const Container = styled.div`
-  border-right: 1px solid #d8dee4;
+import UDropdownMenu from 'components/DropdownMenu';
+import {colors} from 'styles';
+
+export const Bottom = styled.div``;
+
+export const ButtonContainer = styled.div`
   display: flex;
   flex-direction: column;
   gap: 8px;
+`;
+
+export const Container = styled.div`
+  border-right: 1px solid ${colors.border};
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
   overflow-y: auto;
   padding: 24px 16px;
+`;
+
+export const DropdownMenu = styled(UDropdownMenu)`
+  margin-left: 8px;
+`;
+
+export const Top = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
 `;
