+++ src/containers/MainArea/Styles.ts
@@ -1,9 +1,9 @@
 import styled from 'styled-components';
 
-import {hiddenScroll} from 'styles';
+import {colors, hiddenScroll} from 'styles';
 
 export const Container = styled.div`
-  border-right: 1px solid #d8dee4;
+  border-right: 1px solid ${colors.border};
   padding: 8px 0 24px;
   ${hiddenScroll};
 `;
